writer.go: make implement message.Writer interface
The reason for breaking API from prior is that the previous
definition of Write did not implement io.Reader properly.
There is no way for it to implement io.Reader properly.
When io returns EOF, readers are not supposed to read anymore.

This goes against the very definition of a jlog reader.

Also nobody uses this package yet so it's fine (does anybody
use jlog?)

Changing writer to implement message.Writer for consistency.